requirements.txt
anyascii==0.1.7; python_version >= "3.3"
asgiref==3.3.1; python_version >= "3.5"
beautifulsoup4==4.8.2
boto3==1.17.22; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
botocore==1.20.22; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
certifi==2020.12.5
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
dj-database-url==0.5.0
django-basic-auth-ip-whitelist==0.3.4; python_version >= "3.4"
django-crispy-forms==1.11.1; python_version >= "3.5"
django-csp==3.6
django-filter==2.4.0; python_version >= "3.5"
django-formtools==2.2
django-modelcluster==5.1; python_version >= "3.5"
django-pattern-library==0.3.0; python_version >= "3.6" and python_version < "4.0"
django-recaptcha==2.0.6
django-redis==4.11.0; python_version >= "3.5"
django-referrer-policy==1.0
django-sesame==2.3; python_version >= "3.6" and python_version < "4.0"
django-storages==1.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
django-taggit==1.3.0; python_version >= "3.5"
django-treebeard==4.5.1; python_version >= "3.6"
django==3.1.7; python_version >= "3.6"
djangorestframework==3.12.2; python_version >= "3.5"
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
gunicorn==20.0.4; python_version >= "3.4"
html5lib==1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
jdcal==1.4.1
jmespath==0.10.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
l18n==2020.6.1
markdown==3.3.4; python_version >= "3.6"
notifications-python-client==6.0.2; python_version >= "3.6"
openpyxl==3.0.6; python_version >= "3.6"
pillow==8.1.2; python_version >= "3.6"
psutil==5.8.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
psycopg2==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pyjwt==2.0.1; python_version >= "3.6"
python-dateutil==2.8.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pytz==2021.1
pyyaml==5.4.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
raven==6.10.0
redis==3.5.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
requests==2.25.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
s3transfer==0.3.4
scout-apm==2.10.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0" and python_version < "4")
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
soupsieve==2.2; python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.5"
tablib==3.0.0; python_version >= "3.6"
urllib3==1.26.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
wagtail-accessibility==0.2.1
wagtail-autocomplete==0.6
wagtail-django-recaptcha==1.0
wagtail-storages==0.0.6; python_version >= "3.5"
wagtail==2.12.3; python_version >= "3.6"
webencodings==0.5.1
whitenoise==5.0.1; python_version >= "3.5" and python_version < "4"
willow==1.4
wrapt==1.12.1
xlrd==2.0.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
xlsxwriter==1.3.7
xlwt==1.3.0
